additional polish
Moved error codes to one list in game_utils and added function to access them from elsewhere. Changed references in main() to reference as appropriate

Reordered the logic of detecting OS to determine print speed.

Added and Polished up developer testing mode prompts at the beginning of the game, allowing user to pick player items and which descriptions to display, short or regular.

initialized work on an asynchronous input module to allow for user to skip prompts or speed them up as desired. Likely will work on that after submitting project as it's purely optional.